+++ demo/src/main/java/eu/unicate/retroauth/demo/SomeAuthenticatedService.java
@@ -6,13 +6,23 @@ import java.util.List;
 
 import eu.unicate.retroauth.annotations.Authenticated;
 import eu.unicate.retroauth.annotations.Authentication;
+import retrofit.Callback;
 import retrofit.http.GET;
 import retrofit.http.Path;
 import rx.Observable;
 
 @Authentication(accountType = R.string.auth_account_type, tokenType = R.string.auth_token_type)
 public interface SomeAuthenticatedService {
+
 	@Authenticated
 	@GET("/users/{user}/repos")
-	Observable<List<JsonElement>> listRepos(@Path("user") String user);
+	Observable<JsonElement> listReposRxJava(@Path("user") String user);
+
+	@Authenticated
+	@GET("/users/{user}/repos")
+	JsonElement listReposBlocking(@Path("user") String user);
+
+	@Authenticated
+	@GET("/users/{user}/repos")
+	void listReposAsync(@Path("user") String user, Callback<JsonElement> result);
 }