+++ retroauth/src/main/java/eu/unicate/retroauth/AuthRequestType.java
@@ -1,0 +1,16 @@
+package eu.unicate.retroauth;
+
+/**
+ * This enum defines the different request types
+ * handeled by retroauth
+ */
+public enum AuthRequestType {
+	// rxjava calls
+	RXJAVA,
+	// asynchronous calls using the retrofit Callback interface
+	ASYNC,
+	// synchronous calls returning the result object
+	SYNC,
+	// non, means that the original retrofit request will be executed
+	NONE
+}