+++ retroauth/src/test/java/eu/unicate/retroauth/AccountHelperTests.java
@@ -1,0 +1,15 @@
+package eu.unicate.retroauth;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
+
+import static org.junit.Assert.assertTrue;
+
+@RunWith(JUnit4.class)
+public class AccountHelperTests {
+	@Test
+	public void dummyTest() {
+		assertTrue(true);
+	}
+}