+++ retroauth/src/test/java/eu/unicate/retroauth/AllTests.java
@@ -3,8 +3,10 @@ package eu.unicate.retroauth;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
+/**
+ * TODO
+ */
 @RunWith(Suite.class)
-@Suite.SuiteClasses({})
-public class ApplicationTest {
-
+@Suite.SuiteClasses({AccountHelperTests.class})
+public class AllTests {
 }